server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
	"uvicorn[standard]",
	"fastapi[standard]",
	"google-genai>=1.59.0",
	"pydantic>=2.12.5",
	"pillow>=12.0.0",
]
